+++ custom_components/edgeos/clients/__init__.py
@@ -1,0 +1,10 @@
+from homeassistant.exceptions import HomeAssistantError
+
+
+class LoginException(HomeAssistantError):
+    def __init__(self, status_code):
+        self._status_code = status_code
+
+    @property
+    def status_code(self):
+        return self._status_code